Added helpPage feature and updated project configurations
This commit introduces a new feature, the helpPage, with the addition of `helpPage.cpp` and `helpPage.h` files to the project. These files define a new GUI component `aboutPage` with labels, hyperlink buttons, and methods for painting, resizing, and scaling on mobile platforms. The `MainComponent.h` file was updated to include the `helpPage.h` header file and a new shared pointer `pAboutPage`. A new method `checkMicrophoneAccessPermission()` was also added to the MainComponent class.

The `cmp_label.cpp` file was mistakenly removed and then re-added to the project files. The order of compilation for `cmp_label.cpp` and `cmp_legend.cpp` was swapped in the `Makefile` and in the Visual Studio and Xcode project files, possibly for optimization or due to dependency changes.

The Visual Studio Solution File format version was downgraded from 12.00 to 11.00, and several projects and configurations were removed. The project name was changed from "AudioAnalyzer_App" to "AudioAnalyzer - App". The Projucer version used to create some files was updated from 7.0.10 to 7.0.12.