Updated VS Solution File and added new project configurations
Updated the Visual Studio Solution File format version from 11.00 to 12.00 and specified the Visual Studio Version as 17.10.34814.14 with a Minimum Visual Studio Version of 10.0.40219.1. Added a new project "WapProj" to the solution. New configurations for Debug and Release modes have been added for different CPU architectures. Updated configurations for the "AudioAnalyzer_App" project and set the active configuration and build for all architectures to x64. For the "WapProj" project, configurations have been added for all CPU architectures with the active configuration, build, and deploy settings set to their respective architecture. No changes have been made to the SolutionProperties section.