Updated FFTModule methods to manage toggle button states
The `openAudioFile` and `loadURLIntoSpectrum` methods in `FFTModule.h` have been updated to accept two additional `juce::ToggleButton*` parameters. These parameters are pointers to the newly added member variables `spectrumOfaudioFile__toggleButton` and `makespectrumOfInput__toggleButton` in the `FFTModule` class. The state of these buttons is now directly manipulated within these methods. The `timerCallback` method has been updated to reflect the current state of the application in these buttons. In `FFTCtrl.cpp`, the `buttonClicked` method now passes the state of these buttons to the `openAudioFile` method. The lambda function in the `openAudioFile` method now captures and passes these buttons' state to the `loadURLIntoSpectrum` method.