Update FFTModule.h for dynamic timer frequency and image handling
FFTModule.h has been updated with two significant changes. Firstly, the line of code that clears the `spectrogramImage` has been removed, indicating that it no longer needs to be cleared at this point. Secondly, the timer frequency is now dynamically calculated based on the sample rate, number of samples, FFT size, and play time of the audio file, rather than being fixed at 60Hz. This allows for more flexible and accurate timer control based on the properties of the audio file being processed.