Refactor FFTModule and introduce SpectrogramComponent
This commit involves a significant refactoring of the FFTModule and the introduction of a new SpectrogramComponent. The FFTModule.h file was removed from AudioAnalyzer.jucer and replaced with FFTCtrl.cpp, FFTCtrl.h, FFTModule.cpp, and FFTModule.h. These files were later replaced with SpectrogramComponent.cpp and SpectrogramComponent.h. The CMakeLists.txt, AudioAnalyzer.cbp, and Makefile were updated to reflect these changes.

Two new source files FFTCtrl.cpp and FFTModule.cpp were added to the project. They were also included in the AudioAnalyzer\Source filter. A new source file SpectrogramComponent.cpp was added to the project and included in the AudioAnalyzer\Source filter. Two new header files FFTCtrl.h and FFTModule.h were added to the project. A new header file SpectrogramComponent.h was added to the project.

In the FFTCtrl.cpp file, a new toggle button autoSwitchToInput__toggleButton was added. This button allows the user to automatically switch to input. The button's state is also set based on the autoSwitchToInput state of the module_FFT. In the FFTCtrl.h file, a new member variable autoSwitchToInput__toggleButton of type juce::ToggleButton was added. This represents the new toggle button added in the FFTCtrl.cpp file.

The FFTModule.cpp file has been significantly updated with new code. The changes include the addition of a constructor and destructor for the FFTModule class. The implementation of several new methods such as paint, selectFile, handleAudioResource, loadURLIntoFFT, openAudioFile, switchToMicrophoneInput, makeWhiteNoise, makeSines, makeWindows, makeGraphAttributes, showValues, and setAutoSwitchToInput.

The SpectrogramComponent class has been added to SpectrogramComponent.h and SpectrogramComponent.cpp. This class is responsible for managing the audio component of a spectrogram, which is a visual representation of the spectrum of frequencies in a sound or other signal. The OLD_FFTModule.h file has been removed.