Added ScopedLock to ensure thread-safety in SpectrogramComponent
Implemented the use of ScopedLock objects in several methods of the SpectrogramComponent class to ensure thread-safety. The ScopedLock object, named 'sl', is initialized with 'fftLockMutex' and is used to prevent data races in a multithreaded context. The affected methods include 'loadURLIntoSpectrum', 'initRealTimeFftChartPlot', 'fillRTChartPlotFrequencyValues', 'setDoRealTimeFftChartPlot', and 'setMaxFreqInRealTimeFftChartPlot'. These changes ensure that the operations performed in these methods are not executed concurrently by multiple threads.