Refactor SpectrogramComponent to use FFTCtrl and enable real-time FFT
This commit includes a significant refactor of the SpectrogramComponent class. The `std::chrono` library was replaced with a custom `FFTCtrl` class, and the `FFTModule` object is now accessed via a pointer for more flexible management. Two new methods were added: `initRealTimeFftChartPlot()` for initializing the real-time FFT chart plot, and `registerFFTCtrl(FFTCtrl* FFTC)` for registering an `FFTCtrl` object. The `LOG_EXECUTION_TIMES` macro and related logging code were removed, and a check was added to update the sample rate of `ptrFFTCtrl` if it is not null. The `doRealTimeFftChartPlot` member variable was changed to `true`, enabling real-time FFT chart plotting by default. Lastly, a new member variable `ptrFFTCtrl` of type `FFTCtrl*` was added, indicating that an `FFTCtrl` object is now being used within the `SpectrogramComponent` class.