Add notchFiltersCoefficientsMap to NOTCH_50_60_Hz_filter_Coeffs
A new data structure, `notchFiltersCoefficientsMap`, has been added to the `NOTCH_50_60_Hz_filter_Coeffs` namespace. This unordered map is designed to hold integer keys, each associated with another unordered map. The nested unordered map also has integer keys, but each key is associated with a tuple containing two integers and two iterators pointing to elements in a vector of doubles. This change is intended to handle a collection of 50Hz and 60Hz Notch filters coefficients more efficiently. No code was removed in this change.