Add 60Hz Notch filter and reformat coefficient map
The `NOTCH_50_60_Hz_filter_Coeffs.h` file has been updated with two main changes. Firstly, a 60Hz Notch filter at 8000 Hz has been added with its coefficients defined in the `c60HzNotchCoeffsAtFS8000` vector. Secondly, the `notchFiltersCoefficientsMap` unordered map, which stores the coefficients of the 50Hz and 60Hz Notch filters for different frequencies, has been reformatted for better readability. This reformatting includes the addition of line breaks and indentation, but does not alter the functionality or the data stored in the map.