Updated NotchFilter and related components
Significant changes were made to the NotchFilter class and related components. The `NOTCH_50_60_Hz_filter_Coeffs.h` file was updated to correct the coefficients for the 50Hz and 60Hz Notch filters. The `NotchFilter` constructor was modified to accept only `baseFrequency` and `sampleRate` as parameters, and several new methods were added. The `filters` member was replaced with new members related to the notch filter coefficients and input/output values. In `SpectrogramComponent.cpp`, the `loadURLIntoSpectrum` and `makeFilterPing` methods were updated to zero the `fftDataInBuffer` before applying the filter. The `pushNextSampleIntoFifo` method now applies the notch filter to the sample if a filter is selected. The `setFilterToUse` method now creates a `NotchFilter` instead of a `cpEKSNotchFilter`. Lastly, the `theNotchFilter` member in `SpectrogramComponent.h` was updated to be a `std::unique_ptr<NotchFilter>`.