Expand audio file filter, deactivate YLimTimer class
The audio file filter in `AudioPlaybackModule.h` and `FFTModule.h` has been expanded to include `.mp3` files, in addition to the previously supported `.wav`, `.flac`, and `.aif` files. In `SpectrogramComponent.h`, the `YLimTimer` class, which was responsible for calculating the Y limit of the spectrogram component, has been commented out, deactivating its functionality.